adds brown background-color inside cloze
this was added as inside cloze there might be colored text, from bold or
italic text. This would make it hard to see what was filled in by the
cloze hence the added brown background color to make it distinguishable.

at first, I added the background just to the bolded text inside cloze
but it looked wack and hard to distinguish between cloze, bold and
normal text, so the whole cloze now gets the treatment.