adds dfn-tag logic
this is used for japanese words.
this lets japanese words data to be output given that the id
specific data is added for a word. The thinking is
that words data should be output and stored in CSS instead of
copy-pasting between cards
adds grey color to dfn-tags in dialoge dl-tags
adds normal style to dfn tags (AnkiDroid has italic dfn tags while Anki does not)