removes japanese definitions data
this repo had specific css for each defined word. This yielded a css
file with approx 7k lines, mostly for specific words

now that I know the data handling can be done with JS instead (though
that is a hack too as JS does not read from disk). the definitions data
is removed from this repo